+++ Thunderstorm/Modules/Animals/TFishSchools.h
@@ -32,7 +32,7 @@ public:
 	TFishSchools();
 	virtual ~TFishSchools();
 
-	uint32_t ProcessMessage(long _code, MESSAGE & message);
+	uint64_t ProcessMessage(long _code, MESSAGE & message);
 	void Init();
 	void Realize(uint32_t dTime);
 	void Execute(uint32_t dTime);
