+++ Thunderstorm/Modules/Battle_interface/WorldMapInterface/interface.cpp
@@ -58,7 +58,7 @@ void WM_INTERFACE::Realize(uint32_t delta_time)
 	}
 }
 
-uint32_t WM_INTERFACE::ProcessMessage(MESSAGE & message)
+uint64_t WM_INTERFACE::ProcessMessage(MESSAGE & message)
 {
 	switch (message.Long())
 	{
