+++ Thunderstorm/Modules/Battle_interface/image/imgrender.cpp
@@ -28,7 +28,7 @@ void BIImageRender::Render()
 	} while( GetNextPrioritetRange() );
 }
 
-uint32_t BIImageRender::ProcessMessage(MESSAGE & message)
+uint64_t BIImageRender::ProcessMessage(MESSAGE & message)
 {
 	return 0;
 }
