+++ Thunderstorm/Modules/Battle_interface/land/IBoardingStatus.cpp
@@ -27,7 +27,7 @@ bool IBoardingStatus::Init()
 	return true;
 }
 
-uint32_t IBoardingStatus::ProcessMessage(MESSAGE & message)
+uint64_t IBoardingStatus::ProcessMessage(MESSAGE & message)
 {
 	switch (message.Long())
 	{
