+++ Thunderstorm/Modules/Battle_interface/land/battle_land.cpp
@@ -60,7 +60,7 @@ void BATTLE_LAND_INTERFACE::Realize(uint32_t delta_time)
 	}
 }
 
-uint32_t BATTLE_LAND_INTERFACE::ProcessMessage(MESSAGE & message)
+uint64_t BATTLE_LAND_INTERFACE::ProcessMessage(MESSAGE & message)
 {
 	switch (message.Long())
 	{
