+++ Thunderstorm/Modules/Battle_interface/timer/Timer.cpp
@@ -35,7 +35,7 @@ void BITimer::Realize(uint32_t delta_time)
 	}
 }
 
-uint32_t BITimer::ProcessMessage(MESSAGE & message)
+uint64_t BITimer::ProcessMessage(MESSAGE & message)
 {
 	long nMsgCod = message.Long();
 	switch( nMsgCod )
