+++ Thunderstorm/Modules/CameraDialog/CameraDialog.h
@@ -26,7 +26,7 @@ public:
 	bool Init();
 	void Execute(uint32_t Delta_Time);
 	void Realize(uint32_t Delta_Time);
-	uint32_t ProcessMessage(MESSAGE &msg);
+	uint64_t ProcessMessage(MESSAGE &msg);
 	void ProcessStage(Stage stage, uint32_t delta) override
 	{
 		switch (stage)
