+++ Thunderstorm/Modules/Lighter/Lighter.cpp
@@ -159,7 +159,7 @@ void Lighter::Realize(uint32_t delta_time)
 }
 
 //Сообщения
-uint32_t Lighter::ProcessMessage(MESSAGE & message)
+uint64_t Lighter::ProcessMessage(MESSAGE & message)
 {
 	char command[32];
 	message.String(31, command);
