+++ Thunderstorm/Modules/Location/Grass.cpp
@@ -547,7 +547,7 @@ void Grass::Realize(uint32_t delta_time)
 	}
 }
 
-uint32_t Grass::ProcessMessage(MESSAGE &message)
+uint64_t Grass::ProcessMessage(MESSAGE &message)
 {
 	char ctmp[MAX_PATH];
 	switch( message.Long() )
