+++ Thunderstorm/Modules/Location/LocationEffects.h
@@ -98,7 +98,7 @@ public:
 	void Execute(uint32_t delta_time);
 	void Realize(uint32_t delta_time);
 	//Сообщения
-	uint32_t ProcessMessage(MESSAGE & message);
+	uint64_t ProcessMessage(MESSAGE & message);
 	void ProcessStage(Stage stage, uint32_t delta) override
 	{
 		switch (stage)
