+++ Thunderstorm/Modules/Location/ModelRealizer.h
@@ -43,7 +43,7 @@ public:
 	}
 
 	//Сообщения
-	uint32_t ProcessMessage(MESSAGE & message);
+	uint64_t ProcessMessage(MESSAGE & message);
 };
 
 #endif
