+++ Thunderstorm/Modules/Location/WideScreen.cpp
@@ -44,7 +44,7 @@ bool WideScreen::Init()
 }
 
 //Сообщения
-uint32_t WideScreen::ProcessMessage(MESSAGE & message)
+uint64_t WideScreen::ProcessMessage(MESSAGE & message)
 {
 	dlt = -1.0f;
 	return 0;
