+++ Thunderstorm/Modules/Rigging/Flag.h
@@ -66,7 +66,7 @@ public:
 	void Realize(uint32_t Delta_Time);
 	bool CreateState(ENTITY_STATE_GEN * state_gen);
 	bool LoadState(ENTITY_STATE * state);
-    uint32_t ProcessMessage(MESSAGE & message);
+    uint64_t ProcessMessage(MESSAGE & message);
 	void ProcessStage(Stage stage, uint32_t delta) override
 	{
 		switch (stage)
