+++ Thunderstorm/Modules/Sea_ai/SEA_AI.h
@@ -42,7 +42,7 @@ public:
 	bool LoadState(ENTITY_STATE * state);
 	
 	void ProcessMessage(uint32_t iMsg,uint32_t wParam,uint32_t lParam);
-	uint32_t ProcessMessage(MESSAGE & message);
+	uint64_t ProcessMessage(MESSAGE & message);
 
 	uint32_t AttributeChanged(ATTRIBUTES *pAttribute);
 };
