+++ Thunderstorm/Modules/Sea_cameras/SEA_CAMERAS.cpp
@@ -37,7 +37,7 @@ void SEA_CAMERAS::ProcessMessage(uint32_t iMsg,uint32_t wParam,uint32_t lParam)
 	//UNGUARD*/
 }
 
-uint32_t SEA_CAMERAS::ProcessMessage(MESSAGE & message)
+uint64_t SEA_CAMERAS::ProcessMessage(MESSAGE & message)
 {
 	uint32_t i;
 	switch (message.Long())
