+++ Thunderstorm/Modules/Sea_operator/SEA_OPERATOR.h
@@ -41,7 +41,7 @@ public:
 	virtual ~SEA_OPERATOR();
 
 	virtual bool Init();
-	virtual uint32_t ProcessMessage(MESSAGE & message);
+	virtual uint64_t ProcessMessage(MESSAGE & message);
 	virtual void Realize(uint32_t dTime);
 	virtual void Execute(uint32_t dTime);
 	virtual uint32_t AttributeChanged(ATTRIBUTES *_newAttr);
