+++ Thunderstorm/Modules/Seafoam/Seafoam.cpp
@@ -451,7 +451,7 @@ void SEAFOAM::RealizeShipFoam_Mesh(tShipFoamInfo &_shipFoamInfo, uint32_t _dTime
 }
 
 //--------------------------------------------------------------------
-uint32_t SEAFOAM::ProcessMessage(MESSAGE & message)
+uint64_t SEAFOAM::ProcessMessage(MESSAGE & message)
 {
 	//GUARD(SEAFOAM::ProcessMessage)
